requirements.txt
numpy==1.17.4
pandas==0.25.3
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
xlrd==1.2.0
google-api-core==1.8.2
google-auth==1.6.3
google-auth-oauthlib==0.3.0
google-cloud==0.34.0
google-cloud-bigquery==1.11.1
google-cloud-core==0.29.1
google-cloud-logging==1.10.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
